LU-5577 obdclass: cleanup comparison between signed and unsigned
Change arguments for page offset from int to loff_t.
Change type of counter and size variables to size_t.

+++ lustre/include/obd.h
@@ -190,7 +190,7 @@ struct obd_type {
 struct brw_page {
 	obd_off  off;
 	struct page *pg;
-	int count;
+	size_t count;
 	obd_flag flag;
 };
 
@@ -256,7 +256,7 @@ struct client_obd {
 	/* A chunk is an optimal size used by osc_extent to determine
 	 * the extent size. A chunk is max(PAGE_CACHE_SIZE, OST block size) */
 	int			cl_chunkbits;
-	int			cl_extent_tax;	/* extent overhead, by bytes */
+	unsigned int		cl_extent_tax;	/* extent overhead, by bytes */
 
 	/* keep track of objects that have lois that contain pages which
 	 * have been queued for async brw.  this lock also protects the
